requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
deprecation==2.1.0
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.5.0
gotrue==2.12.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
multidict==6.2.0
numpy==2.2.4
packaging==24.2
pluggy==1.5.0
portalocker==2.10.1
postgrest>=0.10.6  # Add explicit postgrest dependency
propcache==0.3.1
protobuf==5.29.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
pytest-order==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2024.1
qdrant-client==1.13.3
realtime==2.4.1
reportlab==4.1.0  # Add this line
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase>=0.7.1,<1.0.0  # Use a stable version
supafunc==0.9.4
tomli==2.2.1
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
yarl==1.18.3

# Add these dependencies
PyPDF2>=3.0.0
nltk>=3.8.1
google-generativeai>=0.3.0
bcrypt>=4.0.1
passlib>=1.7.4
pytest-asyncio>=0.23.0
pytest-timeout>=2.1.0
python-multipart>=0.0.9
email-validator>=2.0.0
